+++ src/schemas/role.schema.js
@@ -1,10 +1,10 @@
 import { z } from "zod";
 
-const thirdPartyAuthSchema = z.object({
+const roleSchema = z.object({
     user_type: z.enum(["applicant","business", "admin"], {
         required_error: "Add user type to query body"
     })
 });
 
 
-export default thirdPartyAuthSchema;
+export default roleSchema;
